+++ packages/amazon-cognito-identity-js/index.d.ts
@@ -368,7 +368,7 @@ declare module 'amazon-cognito-identity-js' {
 		path?: string;
 		expires?: number;
 		secure?: boolean;
-		sameSite?: 'strict' | 'lax';
+		sameSite?: 'strict' | 'lax' | 'none';
 	}
 	export class CookieStorage implements ICognitoStorage {
 		constructor(data: ICookieStorageData);