+++ packages/aws-amplify-react/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.7](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react@4.2.6...aws-amplify-react@4.2.7) (2020-10-01)
+
+**Note:** Version bump only for package aws-amplify-react
+
+
+
+
+
+## [4.2.6](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react@4.2.5...aws-amplify-react@4.2.6) (2020-09-25)
+
+**Note:** Version bump only for package aws-amplify-react
+
+
+
+
+
+## [4.2.5](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react@4.2.4...aws-amplify-react@4.2.5) (2020-09-16)
+
+**Note:** Version bump only for package aws-amplify-react
+
+
+
+
+
+## [4.2.4](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react@4.2.3...aws-amplify-react@4.2.4) (2020-09-15)
+
+**Note:** Version bump only for package aws-amplify-react
+
+
+
+
+
+## [4.2.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react@4.2.2...aws-amplify-react@4.2.3) (2020-09-10)
+
+**Note:** Version bump only for package aws-amplify-react
+
+
+
+
+
+## [4.2.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react@4.2.1...aws-amplify-react@4.2.2) (2020-09-03)
+
+**Note:** Version bump only for package aws-amplify-react
+
+
+
+
+
+## [4.2.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react@4.2.0...aws-amplify-react@4.2.1) (2020-09-03)
+
+**Note:** Version bump only for package aws-amplify-react
+
+
+
+
+
+# [4.2.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react@4.1.23...aws-amplify-react@4.2.0) (2020-09-01)
+
+
+### Features
+
+* **@aws-amplify/ui-components:** Add Chatbot UI to main ([#6684](https://github.com/aws-amplify/amplify-js/issues/6684)) ([4e25f92](https://github.com/aws-amplify/amplify-js/commit/4e25f923723441c6fb705584fcc3349214806558)), closes [#5024](https://github.com/aws-amplify/amplify-js/issues/5024) [#6648](https://github.com/aws-amplify/amplify-js/issues/6648) [#6652](https://github.com/aws-amplify/amplify-js/issues/6652) [#6678](https://github.com/aws-amplify/amplify-js/issues/6678)
+
+
+
+
+
 ## [4.1.23](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react@4.1.22...aws-amplify-react@4.1.23) (2020-08-19)
 
 **Note:** Version bump only for package aws-amplify-react